requirements.txt
discord.py[voice]
pyyaml
requests
GitPython